Add hive pollers to the registery service
After registering the hive with a specific ttl, we need to constantly update the
entry to prevent expiration and also need to watch the directory to know which
hives are joining and leaving the cluster.

This commit adds the functions to watch the registery service and update the
hive's ttl. Also it removes the hive entry from the registery when the hive
stops.