Add transactions and recovery to bee and qee
When a bee calls a receive function or when a qee calls a map function, the
function can panic. In such cases we should recover to make the platform more
robust.

Also for simplicity, the receive function should be transactional. Transactions
should be automatically started. They should be rolled back whenever there is
an error (say a panic) in the application and they should be committed when
there is no error.

Receive functions can now report errors in two forms: They can either panic or
return an error value.