Add raft-based replication to bees
In v0.1, bees used a custom way of replicating their state. With the
new implementation of raft in v0.2, bees should also use raft for the
purpose of replicating transactions. It has more overheads, but the
end-result is exteremely simpler.

This commit adds the first version of that with no support for fault
tolerance and elasticity. Next CLs will add support for those.