Change application options
We relied on specific flags and settings for application options. This
commit switches from that mechanism to generic functional options. As a
result options are passed right when the application is created. And
cannot be changed afterwards. We also dropped support for commit threshold
as it is not required in our new design.