Pace messages based on max bucket size
The number of incoming messages (ie, the batch size) and the number of
emitting messages can be beyond the maximum capacity of their respective
token bucket. This commit limits the batch size and also the number of
outgoing messages based on the in/out bucket max capacity.

Although, it is always recommended to use a max size that would be
well beyond these numbers for a given application.

Change-Id: I23b1f067681f19325cf1ff4c087e4036782ea856