Change the type of value to interface{}
We delegated the task of encoding/decoding values to the applications,
and provided helper functions for encoding/decoding using gob.
For each message, the application had to pay the overhead of
encoding/decoding and that was very significant on batches.
At the end we had to encode every entry using gob raft replication,
which effectively means another significant overhead.

This commit changes the type of dictionary values from []byte to
interface{}. This has two major advantages:
1) Applications will be much simpler.
2) Transactions will be much faster.

Change-Id: Ib3cbadc706dd9e5cafe07e3f7cd3c60e068aaab3