Wait for group election to be finished
This commit adds the capability to wait on election results when the
node leader is not elected yet. This results in much faster retries
when the election process takes less than the retry timeout and also
allows the clients to use a larger timeout without being concerned
about the minimum wait time prior to the first sucessful election.

Change-Id: I97bc5e3a09b6b0456b3b4973b888366c1b620319