Fix deferred reply test
This commit fixes a bug in the deferred reply test. We used to emit
ping messages in reply to unicast and broadcast messages. This can result
in an asynchronous loop and multiple call on closing the done channel.

This commit changes the test in a way that messages are only emitted in
response to unicast messages.

Change-Id: Iebbae2ac0c3f215b5de8a324f02b8c3cd0d20aa0