Fix race for raft logs
Raft has a global logger and if we load multiple hives in the same
program (as we do for test cases) we end up with a race. This commit
applies changes to the raft logger once. The limitation would be that
the logger of the raft library would be set according to the
configuration of the first hive.

Change-Id: Ic2255d659dab223d2fb985c221aaa0f9478bdd11