Fix race in replication process
This commit eliminates the race we have in the replication testcases. In
normal setting (when all messages go through bee's handleMsg) we will
not encounter a race. However, in testcases we might have a race.

This commit locks the bee before starting the replication to make sure
such races will never happen.

Change-Id: I59eb63e943d916b1a11c574fee26522af82ca5f2