Prioritize messages
This commit changes the behavior of the multinode to prioritize
heartbeat messages and send them in a batch. The goal is to prevent
false elections.

On top of that, we allocate one "applier" go-routine for the ndoe,
and one "saver" go-routine per group to save the state for that group
instead of launching a go-routine for every ready.

Change-Id: I84a97f54dc5b742e3e640d95a31b902c99f0cd59