Batch raft proposals if possible
Instead of allowing the proposals to go directly to the underlying
multiraft node, we send them to our raft wrapper and try to batch
them if possible.

This way when we are applying a batch of proposals, we will not
handle readies concurrently, which reduces the number of messages
to send and the size of the state to save.

Change-Id: I65517eb80721004cd12211eafaef80890ba463dc