Add configurable fsync
etcd's WAL implementation, calls fsync on every single write. This might
be needed for some applications but has an exteremely high overhead.
We implemented a fsync time parameter which can be used to configure
the frequency of fsyncs in the WAL. Using a sync time of 0, is identical
to the previous conservative behavior.

Change-Id: I3cc3d40105fd86bb9e1bf10194cbeabb773f72c5