Do not decode local requests
In the raft node, we have to decode requests to apply them in the state
machine. As a result, gob is one of the bottlenecks of the system. To
reduce the load on gob, this commit adds a micro-optimization to
avoid decoding requests that are locally available (made by the same
node in the same run.

Change-Id: I3bbd537b1dd201d5ddaab215e842052ecf3bca3a