Return ErrDuplicateBee only for different info
Registry used to return a duplicate bee error whenever a bee with the
same ID was registered in the registry. Sometimes, because of the
retry semantics, we might observe the same request twice. To fix this
issue, this commit changes the behavior of the registry to generate
error only for different requests.

Change-Id: Id20aaac93205bff56ce05c4f91559e4222be32b3